ref: introduce message module
